[UMPA-256] Add : Service 엔티티에 좋아요 필드 추가
