+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/AccompanistService/AccompanistServiceDetailView.swift
@@ -39,9 +39,8 @@ struct AccompanistServiceDetailView: View {
         .padding(.bottom, ServiceDetailConstant.bottomActionBarHeight)
       }
 
-      BottomActionBar(
-        height: ServiceDetailConstant.bottomActionBarHeight,
-        isLiked: false, // TODO: isLiked 를 받아와야 함
+      ServiceDetailBottomActionBar(
+        isLiked: service.isLiked,
         likeButtonAction: { isLiked in
           serviceDetailInteractor.markAsLike(isLiked, for: service.id)
         },
